requirements.txt
python-dotenv
torch
transformers
scikit-learn
matplotlib
seaborn
pandas
numpy
sentence-transformers
fitz